Cleaning up code

+++ Source/Interactions/Cursor_input.cpp
@@ -10,8 +10,8 @@
 void cursor_position_callback(GLFWwindow* window, double xpos, double ypos)
 {
     // Convert double to string
-    std::string varAsString = std::to_string(xpos);
-    varAsString = varAsString + "  " + std::to_string(ypos);
+    std::string varAsString = std::to_string((int)xpos);
+    varAsString = varAsString + "  " + std::to_string((int)ypos);
 
     Log.info(varAsString);
 }
